requirements.txt
dashscope==1.24.6
langchain==0.3.27
langchain_community==0.3.30
langchain_core==0.3.76
langchain-deepseek==0.1.4
langchain-text-splitters==0.3.11
PyPDF2==3.0.1
python-dotenv==1.1.1
streamlit==1.50.0
